+++ client/src/components/forms/CosmologyForm.vue
@@ -29,8 +29,8 @@
 </template>
 
 <script>
-import BACKEND_CONSTANTS from '../constants/backend_constants';
-import DoubleField from './DoubleField.vue';
+import BACKEND_CONSTANTS from '@/constants/backend_constants';
+import DoubleField from '@/components/DoubleField.vue';
 
 export default {
   name: 'CosmologyForm',
