+++ client/src/components/forms/FilterForm.vue
@@ -36,8 +36,8 @@
 </template>
 
 <script>
-import BACKEND_CONSTANTS from '../constants/backend_constants';
-import DoubleField from './DoubleField.vue';
+import BACKEND_CONSTANTS from '@/constants/backend_constants';
+import DoubleField from '@/components/DoubleField.vue';
 
 const filterChoices = {
   TopHat: 'Top-hat',
