+++ client/src/components/forms/GrowthForm.vue
@@ -30,8 +30,8 @@
 </template>
 
 <script>
-import BACKEND_CONSTANTS from '../constants/backend_constants';
-import DoubleField from './DoubleField.vue';
+import BACKEND_CONSTANTS from '@/constants/backend_constants';
+import DoubleField from '@/components/DoubleField.vue';
 
 const growthChoices = {
   Integral: 'GrowthFactor',
