+++ client/src/components/forms/HMFForm.vue
@@ -41,8 +41,8 @@
 </template>
 
 <script>
-import DoubleField from './DoubleField.vue';
-import BACKEND_CONSTANTS from '../constants/backend_constants';
+import DoubleField from '@/components/DoubleField.vue';
+import BACKEND_CONSTANTS from '@/constants/backend_constants';
 
 const hmfChoices = {
   'Press-Schechter (1974)': 'PS',
