+++ client/src/components/forms/HaloModelForm.vue
@@ -74,9 +74,9 @@
 </template>
 
 <script>
-import BACKEND_CONSTANTS from '../constants/backend_constants';
-import DoubleField from './DoubleField.vue';
-import InputField from './InputField.vue';
+import BACKEND_CONSTANTS from '@/constants/backend_constants';
+import DoubleField from '@/components/DoubleField.vue';
+import InputField from '@/components/InputField.vue';
 
 const haloModelChoices = {
   linear: 'linear',
@@ -170,9 +170,9 @@ export default {
 </script>
 
 <style scoped>
-  @import "../../node_modules/@syncfusion/ej2-base/styles/material.css";
-  @import "../../node_modules/@syncfusion/ej2-buttons/styles/material.css";
-  @import "../../node_modules/@syncfusion/ej2-popups/styles/material.css";
-  @import "../../node_modules/@syncfusion/ej2-inputs/styles/material.css";
+  @import "../../../node_modules/@syncfusion/ej2-base/styles/material.css";
+  @import "../../../node_modules/@syncfusion/ej2-buttons/styles/material.css";
+  @import "../../../node_modules/@syncfusion/ej2-popups/styles/material.css";
+  @import "../../../node_modules/@syncfusion/ej2-inputs/styles/material.css";
 
 </style>
