+++ client/src/components/forms/MassDefinitionForm.vue
@@ -31,8 +31,8 @@
 
 <script>
 
-import BACKEND_CONSTANTS from '../constants/backend_constants';
-import DoubleField from './DoubleField.vue';
+import BACKEND_CONSTANTS from '@/constants/backend_constants';
+import DoubleField from '@/components/DoubleField.vue';
 
 const massDefinitionChoices = {
   'Use native definition of mass function': 'SOGeneric',
