+++ client/src/views/Create.vue
@@ -35,7 +35,7 @@
 <script>
 import INITIAL_STATE from '@/constants/initial_state.json';
 import clonedeep from 'lodash.clonedeep';
-import Forms from '@/components/Forms.vue';
+import Forms from '@/components/forms';
 
 export default {
   name: 'Create',
