+++ client/tests/unit/Create.spec.js
@@ -1,18 +1,17 @@
 import { shallowMount, createLocalVue } from '@vue/test-utils';
 import Create from '@/views/Create.vue';
-import Concentration from '@/components/Concentration.vue';
-import HaloExclusion from '@/components/HaloExclusion.vue';
-import BiasForm from '@/components/BiasForm.vue';
-import HMFForm from '@/components/HMFForm.vue';
-import HODForm from '@/components/HODForm.vue';
-import Profile from '@/components/Profile.vue';
-import ModelMetadataForm from '@/components/ModelMetadataForm.vue';
-import CosmologyForm from '@/components/CosmologyForm.vue';
-import MassDefinitionForm from '@/components/MassDefinitionForm.vue';
-import GrowthForm from '@/components/GrowthForm.vue';
-import HaloModelForm from '@/components/HaloModelForm.vue';
-import FilterForm from '@/components/FilterForm.vue';
-import TransferForm from '@/components/TransferForm.vue';
+import Concentration from '@/components/forms/ConcentrationForm.vue';
+import HaloExclusion from '@/components/forms/HaloExclusionForm.vue';
+import BiasForm from '@/components/forms/BiasForm.vue';
+import HMFForm from '@/components/forms/HMFForm.vue';
+import HODForm from '@/components/forms/HODForm.vue';
+import Profile from '@/components/forms/ProfileForm.vue';
+import CosmologyForm from '@/components/forms/CosmologyForm.vue';
+import MassDefinitionForm from '@/components/forms/MassDefinitionForm.vue';
+import GrowthForm from '@/components/forms/GrowthForm.vue';
+import HaloModelForm from '@/components/forms/HaloModelForm.vue';
+import FilterForm from '@/components/forms/FilterForm.vue';
+import TransferForm from '@/components/forms/TransferForm.vue';
 import constants from '@/constants/initial_state.json';
 import VueMaterial from 'vue-material';
 import VueObserveVisibility from 'vue-observe-visibility';
@@ -49,7 +48,6 @@ describe('Mounted Create', () => {
     expect(wrapper.findComponent(HMFForm)).toBeTruthy();
     expect(wrapper.findComponent(HODForm)).toBeTruthy();
     expect(wrapper.findComponent(MassDefinitionForm)).toBeTruthy();
-    expect(wrapper.findComponent(ModelMetadataForm)).toBeTruthy();
     expect(wrapper.findComponent(TransferForm)).toBeTruthy();
   });
 });
