+++ client/tests/unit/GrowthForm.spec.js
@@ -1,7 +1,7 @@
 /* eslint-disable no-await-in-loop, no-restricted-syntax, no-continue */
 
 import { mount, createLocalVue } from '@vue/test-utils';
-import GrowthForm from '@/components/GrowthForm';
+import GrowthForm from '@/components/forms/GrowthForm';
 import BACKEND_CONSTANTS from '@/constants/backend_constants';
 
 describe('Mounted GrowthForm', () => {
