+++ client/tests/unit/HMFForm.spec.js
@@ -1,7 +1,7 @@
 /* eslint-disable no-await-in-loop, no-restricted-syntax, no-continue */
 
 import { mount, createLocalVue } from '@vue/test-utils';
-import HMFForm from '@/components/HMFForm';
+import HMFForm from '@/components/forms/HMFForm';
 import BACKEND_CONSTANTS from '@/constants/backend_constants';
 
 describe('Mounted HMFForm', () => {
