+++ client/tests/unit/HODForm.spec.js
@@ -1,7 +1,7 @@
 /* eslint-disable no-await-in-loop, no-restricted-syntax, no-continue */
 
 import { mount, createLocalVue } from '@vue/test-utils';
-import HODForm from '@/components/HODForm';
+import HODForm from '@/components/forms/HODForm';
 import BACKEND_CONSTANTS from '@/constants/backend_constants';
 
 describe('Mounted HODForm', () => {
