+++ client/tests/unit/MassDefinitionForm.spec.js
@@ -1,7 +1,7 @@
 /* eslint-disable no-await-in-loop, no-restricted-syntax, no-continue */
 
 import { mount, createLocalVue } from '@vue/test-utils';
-import MassDefinitionForm from '@/components/MassDefinitionForm';
+import MassDefinitionForm from '@/components/forms/MassDefinitionForm';
 import BACKEND_CONSTANTS from '@/constants/backend_constants';
 
 describe('Mounted MassDefinitionForm', () => {
